Connected map controls and scatterpot controls
Sliders should be working with the map and scatterplot both. Clicking a
zip circle on the map should filter the scatterplot data as well.
Remaining problem:  clicking "sold" radio button limits the scatterplot
to under $500k, for some reason. And there are probably other bugs I
haven't yet spotted due to the combinatorics of controls that we have...